only deleting quotes from firebase from AC when we are logged in

+++ src/actions/index.ts
@@ -2,26 +2,30 @@ import { Quote } from "../types";
 
 export const addToFavorites =
   (quote: Quote) => async (dispatch: any, getState: any) => {
-    const { userId } = getState().auth;
-    await fetch(
-      `https://quote-generator-8312a-default-rtdb.firebaseio.com/${userId}/${quote.id}.json`,
-      {
-        method: "PUT",
-        body: JSON.stringify(quote),
-      }
-    );
+    const { userId, isLoggedIn } = getState().auth;
+
+    if (isLoggedIn) {
+      await fetch(
+        `https://quote-generator-8312a-default-rtdb.firebaseio.com/${userId}/${quote.id}.json`,
+        {
+          method: "PUT",
+          body: JSON.stringify(quote),
+        }
+      );
+    }
 
     dispatch({ type: "ADD_TO_FAVORITES", payload: quote });
   };
 
 export const removeFromFavorites =
   (id: number) => async (dispatch: any, getState: any) => {
-    const { userId } = getState().auth;
+    const { userId, isLoggedIn } = getState().auth;
 
-    await fetch(
-      `https://quote-generator-8312a-default-rtdb.firebaseio.com/${userId}/${id}.json`,
-      { method: "DELETE" }
-    );
+    if (isLoggedIn)
+      await fetch(
+        `https://quote-generator-8312a-default-rtdb.firebaseio.com/${userId}/${id}.json`,
+        { method: "DELETE" }
+      );
 
     dispatch({ type: "REMOVE_FROM_FAVORITES", payload: id });
   };
